Add opaque blits.
png files can now be opened. In opaque mode, the pixel colors are
translated when the file is loaded with the proper color, so that the
image can be blitted with memset later.

A second step is to add a transparency feature that can be used to
alpha blit images if needed. This will probably involve a second array
containing the alpha values, so that we can still opaque blit with
memset if we don't care about the transparency.