framebuffer: Introduce the color_t typedef
This is basically an uint32_t wrapped inside a struct. This allows to
prevent bad cast from a raw uint32_t or uint16_t.

color_t variables always correspond to the translated type of the
framebuffer. This means that once we give a color_t we won't have to
wonder whether we need to translate this to the framebuffer native
interface, and we might also add more supported formats more easily if
needed.

Signed-off-by: Guillaume Clement <gclement@baobob.org>